+++ mail-app/src/main/java/maeilmail/BasePackages.java
@@ -1,6 +1,6 @@
 package maeilmail;
 
-class BasePackages {
+public class BasePackages {
 
     public static final String MAEIL_MAIL = "maeilmail";
     public static final String MAEIL_WIKI = "maeilwiki";
