+++ mail-app/src/main/java/maeilmail/MaeilMailApplication.java
@@ -1,5 +1,8 @@
 package maeilmail;
 
+import static maeilmail.BasePackages.MAEIL_MAIL;
+import static maeilmail.BasePackages.MAEIL_WIKI;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
@@ -7,9 +10,9 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableScheduling
-@ConfigurationPropertiesScan(basePackages = {BasePackages.MAEIL_MAIL, BasePackages.MAEIL_WIKI})
+@ConfigurationPropertiesScan(basePackages = {MAEIL_MAIL, MAEIL_WIKI})
 @SpringBootApplication(
-        scanBasePackages = {BasePackages.MAEIL_MAIL, BasePackages.MAEIL_WIKI},
+        scanBasePackages = {MAEIL_MAIL, MAEIL_WIKI},
         exclude = HibernateJpaAutoConfiguration.class
 )
 public class MaeilMailApplication {
