+++ mail-app/src/main/java/maeilmail/config/AsyncConfig.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.config;
 
 import java.util.concurrent.Executor;
 import org.springframework.context.annotation.Bean;
