+++ mail-app/src/main/java/maeilmail/config/CacheConfig.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.config;
 
 import java.util.Arrays;
 import java.util.List;
