+++ mail-app/src/main/java/maeilmail/config/WebMvcConfig.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.config;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
